E-Commerce Data and Customer Retention Analysis
E-Commerce Data and Customer Retention Analysis with SQL
_________________________________________________________
An e-commerce organization demands some analysis of sales and delivery
processes. Thus, the organization hopes to be able to predict more easily the
opportunities and threats for the future.
Acording to this scenario, You are asked to make the following analyzes consistant
with following the instructions given.
Introduction
- You have to create a database and import into the given csv files. (You should
research how to import a .csv file or a .xlsx file)
- During the import process, you will need to adjust the date columns. You need
to carefully observe the data types and how they should be.
- The data are not very clean and fully normalized. However, they don't prevent
you from performing the given tasks. In some cases you may need to use the
string, window, system or date functions.
- There may be some issues where you need to update the tables.
- Manually verify the accuracy of your analysis.